Fix delete button on channel-created links
Phoenix wraps links in a form in order to submit non-get hyperlinks.
Unfortunately elements added after the page has rendered don't get the
even attached to them. This fixes that issue... Would be great for
events to be watched for at the document level. Not sure if that's
doable without a lib such as jquery loaded.